build&fix: enable -Wall -Werror & fix all warnings
- comparison of integer expressions of different signednes
- syntax.y  ISO C++ forbids converting a string constant to ‘char*’ [-Werror=write-strings]
  fix with strdup()

+++ src/constantNode.cpp
@@ -224,7 +224,7 @@ void checkAndComputeConstant(Node* node) {
 
 // compute constant val
 void constantPropagation(graph* g) {
-  for(int i = 0; i < g->sorted.size(); i++) {
+  for(size_t i = 0; i < g->sorted.size(); i++) {
     if(N(i)->status != VALID_NODE) continue;
     totalNode ++;
     switch(N(i)->type) {
@@ -252,7 +252,7 @@ void constantPropagation(graph* g) {
         break;
     }
   }
-  for(int i = 0; i < g->sorted.size(); i++) {
+  for(size_t i = 0; i < g->sorted.size(); i++) {
     if(N(i)->status == VALID_NODE) {
       N(i)->prev.erase(std::remove_if(N(i)->prev.begin(), N(i)->prev.end(),
         [](const Node* n) { return n->status != VALID_NODE; }), N(i)->prev.end());
